try fix dep

+++ Sources/SwiftExtensionsPack/Date/DateCommon.swift
@@ -24,35 +24,35 @@ extension Date {
         return dateWithTimeZone(dateFormatter, secondsFromGMT: secondsFromGMT)
     }
     
-    public func day(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "dd", secondsFromGMT: secondsFromGMT))
+    public func day(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "dd", secondsFromGMT: secondsFromGMT))
     }
     
-    public func month(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "MM", secondsFromGMT: secondsFromGMT))
+    public func month(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "MM", secondsFromGMT: secondsFromGMT))
     }
     
-    public func year(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "yyyy", secondsFromGMT: secondsFromGMT))
+    public func year(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "yyyy", secondsFromGMT: secondsFromGMT))
     }
     
-    public func hours(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "HH", secondsFromGMT: secondsFromGMT))
+    public func hours(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "HH", secondsFromGMT: secondsFromGMT))
     }
     
-    public func minutes(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "mm", secondsFromGMT: secondsFromGMT))
+    public func minutes(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "mm", secondsFromGMT: secondsFromGMT))
     }
     
-    public func seconds(secondsFromGMT: Int?=nil) -> Int16? {
-        return Int16(toString(dateFormat: "ss", secondsFromGMT: secondsFromGMT))
+    public func seconds(secondsFromGMT: Int?=nil) -> UInt? {
+        return UInt(toString(dateFormat: "ss", secondsFromGMT: secondsFromGMT))
     }
 
-    public func toMillis() -> Int64 {
-        return Int64(self.timeIntervalSince1970 * 1000)
+    public func toMillis() -> UInt {
+        return UInt(self.timeIntervalSince1970 * 1000)
     }
 
-    public func toSeconds() -> Int64 {
-        return Int64(self.timeIntervalSince1970)
+    public func toSeconds() -> UInt {
+        return UInt(self.timeIntervalSince1970)
     }
 }
